fix: change generator entry point
